+++ .github/pull_request_template.md
@@ -1,19 +1,9 @@
 ## 무슨 이유로 코드를 변경했는지
 
-
-
 ## 어떤 위험이나 장애가 발견되었는지
-
-
 
 ## 어떤 부분에 리뷰어가 집중하면 좋을지
 
-
-
 ## 관련 스크린샷
 
-
-
 ## 테스트 계획 또는 완료 사항
-
-
