+++ src/components/atoms/Dropdown/CardSortDropdown.jsx
@@ -9,7 +9,8 @@ import useCardSortDropdown from './useCardSortDropdown.hook';
 
 /**
  *
- * @param {CardSort : string | onSortButtonClick : eventHandlerFunction} param0
+ * @param {string} CardSort : 카드 정렬 기준
+ * @param {function} onSortButtonClick : 카드 정렬 기준 변경 시 실행되는 함수
  * @returns
  */
 export default function CardSortDropdown({
