+++ src/components/organisms/PageNationNumbersContainer/PageNationNumbersContainer.jsx
@@ -2,7 +2,10 @@ import PageNationNumbers from '../PageNationNumbers/PageNationNumbers';
 
 /**
  *
- * @param {totalPage : number |  isLoadingUserCardListData : object | onPageClick : eventHandlerFunction |  currentPage : number} param0
+ * @param {number} totalPage : 총 페이지 수
+ * @param {boolean} isLoadingUserCardListData : 유저 카드 리스트 데이터 로딩 상태
+ * @param {function} onPageClick : 페이지네이션 숫자 클릭 시 실행되는 함수
+ * @param {number} currentPage : 현재 페이지
  * @returns 페이지네이션 숫자 목록
  */
