+++ src/components/organisms/RandomQuestionCard/RandomQuestionCard.jsx
@@ -6,12 +6,19 @@ import useRandomQuestionCard from './useRandomQuestionCard.hook';
 /**
  * 실험적인 기능입니다!!
  * @param {onModalCloseButtonClick : event handler, 클릭하면 랜덤 카드가 닫힙니다.} param0
+ * @param {showRandomCard : boolean, 랜덤 카드가 보이는지 안보이는지 여부를 결정합니다.} param1
  * @returns RandomQuestionCard Component
  */
 
-export default function RandomQuestionCard({ onModalCloseButtonClick }) {
+export default function RandomQuestionCard({
+  onModalCloseButtonClick,
+  showRandomCard,
+}) {
   const { randomCard } = useRandomQuestionCard();
 
+  if (!showRandomCard) {
+    return null;
+  }
   return (
     <Wrapper onClick={onModalCloseButtonClick}>
       <CardFlip to={`/post/${randomCard?.id}`}>
