+++ src/constants/page_nation.js
@@ -1,1 +1,3 @@
 export const PAGE_NATION_NUMBER_COUNT = 5;
+
+export const limit = 8;
